+++ CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/fivetran/go-fivetran/compare/v1.0.1...HEAD)
+## [Unreleased](https://github.com/fivetran/go-fivetran/compare/v1.0.2...HEAD)
+
+## [1.0.2](https://github.com/fivetran/go-fivetran/compare/v1.0.1...v1.0.2)
+
+## Added
+Support for custom config (`map[string]interface{}`) for private links:
+- `PrivateLinkCreateService.ConfigCustom` and `PrivateLinkCreateService.DoCustom` methods
+- `PrivateLinkModifyService.ConfigCustom` and `PrivateLinkModifyService.DoCustom` methods
+- `PrivateLinkDetailsService.DoCustom` method
 
 ## [1.0.1](https://github.com/fivetran/go-fivetran/compare/v1.0.0...v1.0.1)